style: rename DeleteZone to DeleteZoneModal
